requirements.txt
python-dotenv==1.1.1
uvicorn==0.35.0
fastapi==0.116.1
pymongo==4.16.0
vapi-server-sdk==1.9.0
tavily-python==0.7.17
openai==1.109.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
googlemaps==4.10.0
phonenumbers==9.0.21